Back-end/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
gunicorn>=20.1,<21.0
psycopg2-binary>=2.9,<3.0
django-storages>=1.13,<2.0
boto3>=1.26,<2.0
Pillow>=9.0,<10.0
